Revert "orderTicket駝峰更新"
