+++ src/main/webapp/food/css/orderStore.css
@@ -1,10 +1,5 @@
 @import url('https://fonts.googleapis.com/css?family=Montserrat');
-h1 {
-	font-weight: bolder;
-	text-align: center;
-	color: white;
-	font-size: 3em;
-}
+
 h2 {
 	font-weight: bolder;
 	text-align: center;
